+++ crates/tauri-bundler/src/bundle/windows/msi/wix.rs
@@ -304,11 +304,7 @@ fn run_candle(
     }
   };
 
-  let main_binary = settings
-    .binaries()
-    .iter()
-    .find(|bin| bin.main())
-    .ok_or_else(|| anyhow::anyhow!("Failed to get main binary"))?;
+  let main_binary = settings.main_binary()?;
 
   let mut args = vec![
     "-arch".to_string(),
@@ -403,13 +399,6 @@ pub fn build_wix_app_installer(
 
   // target only supports x64.
   log::info!("Target: {}", arch);
-
-  let main_binary = settings
-    .binaries()
-    .iter()
-    .find(|bin| bin.main())
-    .ok_or_else(|| anyhow::anyhow!("Failed to get main binary"))?;
-  let app_exe_source = settings.binary_path(main_binary);
 
   let output_path = settings.project_out_directory().join("wix").join(arch);
 
@@ -524,11 +513,17 @@ pub fn build_wix_app_installer(
   data.insert("manufacturer", to_json(manufacturer));
   let upgrade_code = Uuid::new_v5(
     &Uuid::NAMESPACE_DNS,
-    format!("{}.app.x64", &settings.main_binary_name()).as_bytes(),
+    format!("{}.exe.app.x64", &settings.product_name()).as_bytes(),
   )
   .to_string();
 
   data.insert("upgrade_code", to_json(upgrade_code.as_str()));
+  let product_code = Uuid::new_v5(
+    &Uuid::NAMESPACE_DNS,
+    &settings.bundle_identifier().as_bytes(),
+  )
+  .to_string();
+  data.insert("product_code", to_json(product_code.as_str()));
   data.insert(
     "allow_downgrades",
     to_json(settings.windows().allow_downgrades),
@@ -539,9 +534,6 @@ pub fn build_wix_app_installer(
 
   let shortcut_guid = generate_package_guid(settings).to_string();
   data.insert("shortcut_guid", to_json(shortcut_guid.as_str()));
-
-  let app_exe_name = settings.main_binary_name().to_string();
-  data.insert("app_exe_name", to_json(app_exe_name));
 
   let binaries = generate_binaries_data(settings)?;
 
@@ -565,7 +557,13 @@ pub fn build_wix_app_installer(
   let merge_modules = get_merge_modules(settings)?;
   data.insert("merge_modules", to_json(merge_modules));
 
-  data.insert("app_exe_source", to_json(app_exe_source));
+  // Note: `main_binary_name` is not used in our template but we keep it as it is potentially useful for custom temples
+  let main_binary_name = settings.main_binary_name()?;
+  data.insert("main_binary_name", to_json(main_binary_name));
+
+  let main_binary = settings.main_binary()?;
+  let main_binary_path = settings.binary_path(main_binary);
+  data.insert("main_binary_path", to_json(main_binary_path));
 
   // copy icon from `settings.windows().icon_path` folder to resource folder near msi
   let icon_path = copy_icon(settings, "icon.ico", &settings.windows().icon_path)?;