+++ crates/tauri-cli/src/mobile/ios/build.rs
@@ -174,6 +174,11 @@ pub fn command(options: Options, noise_level: NoiseLevel) -> Result<()> {
   )?;
   inject_resources(&config, tauri_config.lock().unwrap().as_ref().unwrap())?;
 
+  let mut plist = plist::Dictionary::new();
+  let version = interface.app_settings().get_package_settings().version;
+  plist.insert("CFBundleShortVersionString".into(), version.clone().into());
+  plist.insert("CFBundleVersion".into(), version.into());
+
   let info_plist_path = config
     .project_dir()
     .join(config.scheme())
@@ -182,6 +187,7 @@ pub fn command(options: Options, noise_level: NoiseLevel) -> Result<()> {
     info_plist_path.clone().into(),
     tauri_path.join("Info.plist").into(),
     tauri_path.join("Info.ios.plist").into(),
+    plist::Value::Dictionary(plist).into(),
   ])?;
   merged_info_plist.to_file_xml(&info_plist_path)?;
 
@@ -265,12 +271,11 @@ fn run_build(
   crate::build::setup(&interface, &mut build_options, tauri_config.clone(), true)?;
 
   let app_settings = interface.app_settings();
-  let bin_path = app_settings.app_binary_path(&InterfaceOptions {
+  let out_dir = app_settings.out_dir(&InterfaceOptions {
     debug: build_options.debug,
     target: build_options.target.clone(),
     ..Default::default()
   })?;
-  let out_dir = bin_path.parent().unwrap();
   let _lock = flock::open_rw(out_dir.join("lock").with_extension("ios"), "iOS")?;
 
   let cli_options = CliOptions {
@@ -308,13 +313,7 @@ fn run_build(
         build_config = build_config.authentication_credentials(credentials.clone());
       }
 
-      target.build(
-        config,
-        env,
-        NoiseLevel::FranklyQuitePedantic,
-        profile,
-        build_config,
-      )?;
+      target.build(config, env, noise_level, profile, build_config)?;
 
       target.archive(
         config,