+++ crates/tauri-utils/src/platform.rs
@@ -312,7 +312,7 @@ fn resource_dir_from<P: AsRef<std::path::Path>>(
     res = if curr_dir.ends_with("/data/usr/bin") {
       // running from the deb bundle dir
       exe_dir
-        .join(format!("../lib/{}", package_info.crate_name))
+        .join(format!("../lib/{}", package_info.name))
         .canonicalize()
         .map_err(Into::into)
     } else if let Some(appdir) = &env.appdir {
@@ -320,14 +320,11 @@ fn resource_dir_from<P: AsRef<std::path::Path>>(
       Ok(PathBuf::from(format!(
         "{}/usr/lib/{}",
         appdir.display(),
-        package_info.crate_name
+        package_info.name
       )))
     } else {
       // running bundle
-      Ok(PathBuf::from(format!(
-        "/usr/lib/{}",
-        package_info.crate_name
-      )))
+      Ok(PathBuf::from(format!("/usr/lib/{}", package_info.name)))
     };
   }