+++ crates/tauri-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## \[2.0.0-rc.11]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-rc.11`
+
+## \[2.0.0-rc.10]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-rc.10`
+
 ## \[2.0.0-rc.9]
 
 ### Dependencies